requirements.txt
ultralytics
opencv-python-headless
beautifulsoup4
Flask
pyngrok
requests
numpy
Pillow
